Fix: Don't try to delete item from WebFlow when we don't have webflow_item_id or webflow_site_id

+++ app/jobs/webflow_sync/destroy_item_job.rb
@@ -3,6 +3,9 @@
 module WebflowSync
   class DestroyItemJob < ApplicationJob
     def perform(collection_slug:, webflow_site_id:, webflow_item_id:)
+      return if webflow_site_id.blank?
+      return if webflow_item_id.blank?
+
       WebflowSync::Api.new(webflow_site_id).delete_item(collection_slug, webflow_item_id)
     end
   end
